requirements.txt
Flask==2.3.2
Werkzeug==2.3.6
openai-whisper==20230314
pydub==0.25.1
numpy==1.24.3
torch==2.0.1
